+++ my-app/app/team_schedules/_icons/CollapseIcon.tsx
