+++ my-app/app/team_schedules/_icons/SettingsIcon.tsx
