+++ my-app/app/team_schedules/_icons/StatusIcon.tsx
